requirements.txt
blend-modes==2.1.0
cycler==0.11.0
fonttools==4.37.1
imageio==2.21.2
imgaug==0.4.0
kiwisolver==1.4.4
lmdb==1.3.0
matplotlib==3.5.3
networkx==2.8.6
numpy==1.23.2
opencv-python==4.5.4.60
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
scikit-image==0.19.3
scipy==1.9.1
Shapely==1.8.4
six==1.16.0
tifffile==2022.8.12
tqdm==4.64.1
typing-extensions==4.3.0
Wand==0.6.10
